requirements.txt
streamlit==1.38.0
ultralytics==8.3.30
pillow
opencv-python-headless
numpy
torch
torchvision